feat: implement meter API for meter profiling
In this commit, the meter API has been utilized for meter profiling. The measurement of recovery time was removed from the ProfilerEventRecorder and instead the RecoveryMeasurer meter was used. Additional changes include the cleanup setup for the meter and ProfilerExtension, the removal of profile_recovery_time methods in ProfilerWrapper and decorators.py, as well as corresponding changes in test files to reflect these updates. The changes increase the modularity of the code and encapsulate the measurement logic within the Meter classes. This makes it easier to add more complex logic in the future if required.